Initial commit of an AND kernel
Awful name, I know. You provide two input files. For each point in the first
file, look for the nearest neighbor in the second. If it is within the
specified tolerance, tag the point with a 1 in the newly created dimension
(user must provide dimension name). This is written out to the third file.

The output file will have the same number of points as the first file, with a
new dimension that is filled with 0's and 1's.